Add error checks

+++ northslope-setup.sh
@@ -24,5 +24,11 @@ fi
 # Create a temp setup script so that we can pass in command line args
 setup_script=`mktemp`
 curl -fsSL https://raw.githubusercontent.com/northslopetech/setup/refs/heads/latest/setup.sh > ${setup_script}
+if [[ $? -ne 0 ]]; then
+    echo "Error: Unable to download the northslope setup script."
+    exit 1
+fi
 /bin/zsh ${setup_script} "$@"
+exit_code=$?
 rm ${setup_script}
+exit ${exit_code}
